create monthSlider component and simplify monthBar component
